Updates routes to hot reload
